Convert existing translators to lookups
